chore: Move extended storages to a sub package
