+++ ors-api/src/main/java/org/heigit/ors/api/config/profile/ProfileProperties.java
@@ -5,9 +5,10 @@ import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import org.apache.commons.lang3.StringUtils;
 import org.heigit.ors.api.config.EncoderOptionsProperties;
 import org.heigit.ors.api.config.ExecutionProperties;
+import org.heigit.ors.api.config.PreparationProperties;
 import org.heigit.ors.api.config.NonEmptyObjectFilter;
 import org.heigit.ors.api.config.ExtendedStorage;
-import org.heigit.ors.api.config.PreparationProperties;
+import org.heigit.ors.api.config.profile.storages.ExtendedStorage;
 import org.heigit.ors.api.converters.ExtendedStorageMapDeserializer;
 
 import java.nio.file.Paths;
