+++ ors-api/src/main/java/org/heigit/ors/api/config/profile/storages/ExtendedStorageShadowIndex.java
@@ -1,4 +1,4 @@
-package org.heigit.ors.api.config;
+package org.heigit.ors.api.config.profile.storages;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonTypeName;
