+++ ors-api/src/test/java/org/heigit/ors/api/config/ExtendedStorageWayCategoryTest.java
@@ -1,6 +1,8 @@
 package org.heigit.ors.api.config;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+import org.heigit.ors.api.config.profile.storages.ExtendedStorage;
+import org.heigit.ors.api.config.profile.storages.ExtendedStorageWayCategory;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertNotNull;
