+++ ors-api/src/test/java/org/heigit/ors/api/converters/ExtendedStorageMapDeserializerTest.java
@@ -8,7 +8,7 @@ import com.fasterxml.jackson.databind.JsonMappingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.fasterxml.jackson.databind.module.SimpleModule;
-import org.heigit.ors.api.config.ExtendedStorage;
+import org.heigit.ors.api.config.profile.storages.ExtendedStorage;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
